Fixed Courier email processing timeout
EmailProcessor.php now calls Email.php asynchronously and delegates the
work to that class.  This allows EmailProcessor.php to return quickly
and remove the email from the mailq.  If EmailProcessor.php takes too
long, the system will kill the Courier process and restart processing on
the mailq, which would resent the emails from scratch.